+++ connection/src/main/java/cli/user/Administrator.java
@@ -1,7 +1,7 @@
 package cli.user;
 
 public class Administrator extends AbstractUser {
-    public Administrator(String first_name, String last_name, String email, String password, int telephone) {
+    public Administrator(String first_name, String last_name, String email, String password, long telephone) {
         super(first_name, last_name, email, password, telephone,-1,"","");
     }
 }
