+++ connection/src/main/java/cli/user/EvaluatedOrganizationUser.java
@@ -5,11 +5,11 @@ import cli.organization.EvaluatedOrganization;
 public class EvaluatedOrganizationUser extends OrganizationUser {
 
 
-    public EvaluatedOrganizationUser(String first_name, String last_name, String email, String passwordUser, int telephone, EvaluatedOrganization organization) {
+    public EvaluatedOrganizationUser(String first_name, String last_name, String email, String passwordUser, long telephone, EvaluatedOrganization organization) {
         super(first_name, last_name, email, passwordUser, telephone, organization);
     }
 
-    public EvaluatedOrganizationUser(String first_name, String last_name, String email, String passwordUser, int telephone, int idOrganization, String organizationType, String illness) {
+    public EvaluatedOrganizationUser(String first_name, String last_name, String email, String passwordUser, long telephone, int idOrganization, String organizationType, String illness) {
         super(first_name, last_name, email, passwordUser, telephone, idOrganization, organizationType,illness);
     }
 }
